feat: lite avatar handler support multi session
