+++ src/chat_engine/common/handler_base.py
@@ -84,3 +84,6 @@ class HandlerBase(ABC):
     @abstractmethod
     def destroy_context(self, context: HandlerContext):
         pass
+
+    def destroy(self):
+        pass
