+++ src/chat_engine/data_models/chat_engine_config_data.py
@@ -9,6 +9,7 @@ from chat_engine.common.engine_channel_type import EngineChannelType
 class HandlerBaseConfigModel(BaseModel):
     enabled: bool = Field(default=True)
     module: Optional[str] = Field(default=None)
+    concurrent_limit: int = Field(default=1)
 
 
 class ChatEngineOutputSource(BaseModel):
@@ -18,6 +19,7 @@ class ChatEngineOutputSource(BaseModel):
 
 class ChatEngineConfigModel(BaseModel):
     model_root: str = ""
+    concurrent_limit: int = Field(default=1)
     handler_search_path: List[str] = Field(default_factory=list)
     handler_configs: Optional[Dict[str, Dict]] = None
     outputs: Dict[EngineChannelType, ChatEngineOutputSource] = Field(default_factory=dict)
