+++ src/handlers/client/rtc_client/client_handler_rtc.py
@@ -100,7 +100,6 @@ class RtcClientSessionDelegate(ClientSessionDelegate):
 
 class ClientRtcConfigModel(HandlerBaseConfigModel, BaseModel):
     connection_ttl: int = Field(default=900)
-    concurrent_limit: int = Field(default=1)
     turn_config: Optional[Dict] = Field(default=None)
